fixed backlash scaling calculation

+++ src/firmware/end_point_select.c
@@ -63,12 +63,12 @@ static void render_position(struct Bitmap* bm, const struct MotorControl* motor_
 
 static void render_backlash(struct Bitmap* bm, const struct MotorControl* motor_snap) {
   char str[32];
-  const int16_t backlash = (int16_t)((motor_snap->motor_pos / 2) - motor_snap->current_pos);
+  const int32_t backlash = (int16_t)((motor_snap->motor_pos / 2) - motor_snap->current_pos);
   snprintf(str, sizeof(str), "Backlash:  %5d", backlash);
   bitmap_str(bm, terminus8x16, str, 0, 96, bitmap_SET);
 
   // convert backlash from a full scale to a -63 to 63 scale
-  const int16_t scaled_bl = backlash * 63 / motor_snap->backlash;
+  const int32_t scaled_bl = backlash * 63 / (int32_t)motor_snap->backlash;
   bitmap_vline(bm, 0, 120, 7, bitmap_SET);
   bitmap_vline(bm, 32, 122, 5, bitmap_SET);
   bitmap_vline(bm, 64, 120, 7, bitmap_SET);
